functions/create_build_entity_func/requirements.txt
google-cloud-datastore == 1.8.0
